refactor(query): consolidate metric filter validation and extract mongo guard translator
Promote the scalar metric filter walker to a single shared helper in
wow-query (`MetricFilterValidation.requireScalarMetricFilterFields`) so
MongoDB and Elasticsearch enforce one contract from one place, with
direct unit tests in wow-query.

Element-match and search filters are now rejected upfront by the shared
walker with backend-neutral messages, which makes the Mongo translator's
`$elemMatch` branch unreachable - pruned per the reachability doctrine.
Contains-all keeps its per-backend behavior (Elasticsearch terms filter
vs MongoDB untranslatable `$all`).

Extract the MongoDB match-document-to-guard-condition translator into
`MongoMetricFilterGuards.kt` so the compiler file carries only pipeline
assembly.
