+++ wow-elasticsearch/src/main/kotlin/me/ahoo/wow/elasticsearch/query/aggregation/ElasticsearchAggregationCompiler.kt
@@ -21,8 +21,6 @@ import co.elastic.clients.elasticsearch._types.mapping.RuntimeFieldType
 import co.elastic.clients.elasticsearch._types.query_dsl.Query
 import co.elastic.clients.json.JsonData
 import co.elastic.clients.util.NamedValue
-import me.ahoo.wow.api.query.AggregateIdFilter
-import me.ahoo.wow.api.query.AggregateIdsFilter
 import me.ahoo.wow.api.query.AggregationDateUnit
 import me.ahoo.wow.api.query.AggregationExpression
 import me.ahoo.wow.api.query.AggregationExpressionOperator
@@ -30,43 +28,9 @@ import me.ahoo.wow.api.query.AggregationFunction
 import me.ahoo.wow.api.query.AggregationGroup
 import me.ahoo.wow.api.query.AggregationMetric
 import me.ahoo.wow.api.query.AggregationQuery
-import me.ahoo.wow.api.query.AndFilter
-import me.ahoo.wow.api.query.BetweenFilter
-import me.ahoo.wow.api.query.ContainsAllFilter
-import me.ahoo.wow.api.query.ContainsFilter
-import me.ahoo.wow.api.query.DeletionFilter
-import me.ahoo.wow.api.query.ElementMatchFilter
-import me.ahoo.wow.api.query.EndsWithFilter
-import me.ahoo.wow.api.query.EqualFilter
-import me.ahoo.wow.api.query.ExistsFilter
-import me.ahoo.wow.api.query.FilterExpression
-import me.ahoo.wow.api.query.GreaterThanFilter
-import me.ahoo.wow.api.query.GreaterThanOrEqualFilter
-import me.ahoo.wow.api.query.IdFilter
-import me.ahoo.wow.api.query.IdsFilter
-import me.ahoo.wow.api.query.InFilter
-import me.ahoo.wow.api.query.IsEmptyFilter
-import me.ahoo.wow.api.query.IsEmptyStringFilter
-import me.ahoo.wow.api.query.IsNotEmptyStringFilter
-import me.ahoo.wow.api.query.IsNotNullFilter
-import me.ahoo.wow.api.query.IsNullFilter
-import me.ahoo.wow.api.query.LessThanFilter
-import me.ahoo.wow.api.query.LessThanOrEqualFilter
 import me.ahoo.wow.api.query.MatchAllFilter
-import me.ahoo.wow.api.query.MatchNoneFilter
-import me.ahoo.wow.api.query.NorFilter
-import me.ahoo.wow.api.query.NotEqualFilter
-import me.ahoo.wow.api.query.NotExistsFilter
-import me.ahoo.wow.api.query.NotInFilter
-import me.ahoo.wow.api.query.OrFilter
-import me.ahoo.wow.api.query.OwnerIdFilter
 import me.ahoo.wow.api.query.QueryField
-import me.ahoo.wow.api.query.RelativeTimeFilter
-import me.ahoo.wow.api.query.SearchFilter
 import me.ahoo.wow.api.query.Sort
-import me.ahoo.wow.api.query.SpaceIdFilter
-import me.ahoo.wow.api.query.StartsWithFilter
-import me.ahoo.wow.api.query.TenantIdFilter
 import me.ahoo.wow.api.query.schema.QueryCapability
 import me.ahoo.wow.api.query.schema.QueryCardinality
 import me.ahoo.wow.api.query.schema.QueryValueKind
@@ -78,6 +42,7 @@ import me.ahoo.wow.query.schema.QuerySchemaValidationException
 import me.ahoo.wow.query.schema.QueryValueSchema
 import me.ahoo.wow.query.schema.distinctCountCapability
 import me.ahoo.wow.query.schema.physicalField
+import me.ahoo.wow.query.schema.requireScalarMetricFilterFields
 import java.time.Instant
 import java.util.concurrent.TimeUnit
 
@@ -559,66 +524,3 @@ internal class ElasticsearchAggregationCompiler(
             TimeUnit.DAYS -> 86_400_000L to 1L
         }
 }
-
-/**
- * Filter aggregations keep Elasticsearch's element-matching semantics over array fields while
- * MongoDB metric guards compare whole values, so array-valued fields, [ElementMatchFilter],
- * and [SearchFilter] are rejected to mirror the MongoDB compiler's scalar-only metric filter contract.
- */
-@Suppress("CyclomaticComplexMethod", "LongMethod")
-private fun FilterExpression.requireScalarMetricFilterFields(
-    parent: QueryField?,
-    schema: QueryModelSchema,
-) {
-    when (this) {
-        MatchAllFilter, MatchNoneFilter,
-        is IdFilter, is IdsFilter, is AggregateIdFilter, is AggregateIdsFilter,
-        is TenantIdFilter, is OwnerIdFilter, is SpaceIdFilter, is DeletionFilter,
-        -> Unit
-
-        is AndFilter -> operands.forEach { it.requireScalarMetricFilterFields(parent, schema) }
-        is OrFilter -> operands.forEach { it.requireScalarMetricFilterFields(parent, schema) }
-        is NorFilter -> operands.forEach { it.requireScalarMetricFilterFields(parent, schema) }
-        is ElementMatchFilter -> throw QuerySchemaValidationException(
-            "Elasticsearch metric filters cannot translate [ELEMENT_MATCH] into a filter aggregation.",
-        )
-        is SearchFilter -> throw QuerySchemaValidationException(
-            "Aggregation metric filters do not support search filters.",
-        )
-        is RelativeTimeFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is EqualFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is NotEqualFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is GreaterThanFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is GreaterThanOrEqualFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is LessThanFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is LessThanOrEqualFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is BetweenFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is ContainsFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is StartsWithFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is EndsWithFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is InFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is NotInFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is ContainsAllFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is IsEmptyFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is IsEmptyStringFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is IsNotEmptyStringFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is IsNullFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is IsNotNullFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is ExistsFilter -> field.requireScalarMetricFilterField(parent, schema)
-        is NotExistsFilter -> field.requireScalarMetricFilterField(parent, schema)
-    }
-}
-
-private fun QueryField.requireScalarMetricFilterField(parent: QueryField?, schema: QueryModelSchema) {
-    val logical = parent?.append(this) ?: this
-    val value = schema.field(logical)?.value ?: return
-    if (value.isArrayValued) {
-        throw QuerySchemaValidationException(
-            "Aggregation metric filter field [$logical] must be scalar; array fields are not supported in metric filters.",
-        )
-    }
-}
-
-private val QueryValueSchema.isArrayValued: Boolean
-    get() = kind == QueryValueKind.ARRAY ||
-        (kind == QueryValueKind.UNION && alternatives.any { it.kind == QueryValueKind.ARRAY })
